ref: move sdk code from cli
